Renaming constants files, removing login1
