+++ react-native/react/constants/devices.js
