--- react-native/react/constants/login-action-types.js
@@ -1,17 +1,0 @@
-'use strict'
-/* @flow */
-
-export const START_LOGIN = 'START_LOGIN'
-export const LOGGED_IN = 'LOGGED_IN'
-
-export const ASK_USER_PASS = 'ASK_USER_PASS'
-export const SUBMIT_USER_PASS = 'SUBMIT_USER_PASS'
-
-export const ASK_DEVICE_NAME = 'ASK_DEVICE_NAME'
-export const SUBMIT_DEVICE_NAME = 'SUBMIT_DEVICE_NAME'
-export const DEVICE_NAME_TAKEN = 'DEVICE_NAME_TAKEN'
-
-export const ASK_DEVICE_SIGNER = 'ASK_DEVICE_SIGNER'
-export const SUBMIT_DEVICE_SIGNER = 'SUBMIT_DEVICE_SIGNER'
-
-export const SHOW_SECRET_WORDS = 'SHOW_SECRET_WORDS'
