--- react-native/react/constants/login-states.js
@@ -1,8 +1,0 @@
-'use strict'
-/* @flow */
-
-export const ASK_USER_PASS = 'ASK_USER_PASS'
-export const ASK_DEVICE_NAME = 'ASK_DEVICE_NAME'
-export const ASK_DEVICE_SIGNER = 'ASK_DEVICE_SIGNER'
-export const SHOW_SECRET_WORDS = 'SHOW_SECRET_WORDS'
-export const LOGGED_IN = 'LOGGED_IN'
