+++ react-native/react/constants/reducer.js
