+++ react-native/react/constants/search.js
