+++ react-native/react/constants/tabbed-router.js
