+++ react-native/react/reducers/serialize.desktop.js
@@ -1,7 +1,7 @@
 'use strict'
 /* @flow */
 
-import type { State } from '../constants/reducer-types'
+import type { State } from '../constants/reducer'
 
 export default function (state: State, action: any): State {
   return state
